Rework tracing for blocks running as methods
The main impetus for this change is to fix [Bug #13392]. Before this
change, we fired the "return" TracePoint event after popping the stack
frame for the block running as method (BMETHOD). This gave undesirable
source location outputs as the return event normally fires right before
the frame going away.

This change makes vm_trace() fire call/return events for instructions
that have b_call/b_return events attached when the iseq is running as a
BMETHOD. The logic for rewriting to "trace_*" instruction is tweaked so
that when the user listens to call/return events, instructions with
b_call/b_return become trace variants.

To continue to provide the return value for non-local returns done using
the "return" or "break" keyword inside BMETHODs, the stack unwinding
code is tweaked. b_return events now provide the same return value as
return events for these non-local cases. A pre-existing test deemed not
providing a return value for these b_return events as a limitation.

This commit removes the checks for call/return TracePoint events that
happen when calling into BMETHODs when no TracePoints are active.
Technically, migrating just the return event is enough to fix the bug,
but migrating both call and return removes our reliance on
VM_FRAME_FLAG_FINISH and re-entering the interpreter. This change makes
it possible to stay in the interpreter when calling from the interpreter
in the future.