Fix method lookup for calling initialize_copy on live modules
When a module is included, we create an iclass with the same method
table pointer as the original. The old code for Module#initialize_copy
changes the table pointer on the original which makes the lookups on
existing iclasses use the old, outdated table.

Instead of allocating a new table, empty the original table instead.

[Bug #17048]