Fix reinitializing a module with origin using one without
Consider `A.send(:initialize_copy, B)` where A has prepended modules and
B doesn't have any prepended modules. In this case, A has an origin,
but B doesn't.

If A is included by other modules, we have iclasses that share the
method table with A's origin. All these iclasses rely on A to keep the
origin alive. The old code removed A's origin, which led to
use-after-free.

[Bug #17048]