YJIT: ZJIT: Share identical glue functions
Working towards have YJIT and ZJIT in the same build, we need to
deduplicate some glue code that otherwise cause name collision.
Add jit_glue.c for this and build it for YJIT and ZJIT builds.
Update bindgen to look at jit_glue.c; some shuffling of functions
in the output, but the set of functions shouldn't have changed.